Add files via upload
Asked Manus A.I. to build (develop) a WordPress plugin for me. Because I don't know how to program.
But it just don't work.
If someone be interested in go further with this development, I appreciate. 
If it works, of course, I will give the credit whose deserve it.
The idea is to share this plugin freely for the community.